Refactor out details for each photo day from list
